Sistema/segundo_fluxo/services/notification-service/requirements.txt
fastapi[standard]
aio-pika
telegram
python-telegram-bot
pydantic
pymongo
python-dotenv